Change database cause not possibility to salt password with bcrypt that there are already in db

+++ src/main/resources/db/migration/V4__Add_subscriptions_table.sql
@@ -1,5 +1,5 @@
 create table account_subscriptions (
-  subscriber_id bigint not null references account,
-  channel_id bigint not null references account,
+  subscriber_id int8 not null references account,
+  channel_id int8 not null references account,
   primary key (channel_id, subscriber_id)
 )
